product_mangnerger update

+++ src/main/java/com/quynhtd/source_code_final/service/ProductService.java
@@ -2,6 +2,7 @@ package com.quynhtd.source_code_final.service;
 
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,15 +19,27 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Product getProductById(Long id) {
-        return productRepository.findById(id).orElse(null);
+    public Optional<Product> getProductById(Long id) {
+        return productRepository.findById(id);
     }
 
-    public Product saveProduct(Product product) {
+    public Product createProduct(Product product) {
         return productRepository.save(product);
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public Product updateProduct(Long id, Product product) {
+        if (productRepository.existsById(id)) {
+        	product.setId(id);
+            return productRepository.save(product);
+        }
+        return null;
+    }
+
+    public boolean deleteProduct(Long id) {
+        if (productRepository.existsById(id)) {
+        	productRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
